camera orbit, 32bit indices, etc
